+++ lib/src/exceptions/oauth2_protocol_error_exception.dart
@@ -2,16 +2,14 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:ac_dart_essentials/ac_dart_essentials.dart';
-
 import '../oauth2_protocol_error.dart';
+import 'oauth2_http_exception.dart';
 
-class OAuth2ProtocolErrorException extends HttpStatusException {
+class OAuth2ProtocolErrorException extends OAuth2HttpException {
   final OAuth2ProtocolError protocolError;
 
   OAuth2ProtocolErrorException({
     required super.status,
-    super.statusText,
     required super.url,
     required this.protocolError,
   });