+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+- *BREAKING CHANGE*: remove need to specify statusText in OAuth2HttpResponse
+- add OAuth2HttpException instead of using dart:io HttpException
+
 ## 0.1.0
 
 - Initial release